task14: wrong assignment for loop and loop2

+++ 0x13-more_singly_linked_lists/102-free_listint_safe.c
@@ -10,7 +10,7 @@ size_t free_listint_safe(listint_t **h)
 {
 	listint_t *temp;
 	size_t n = 0;
-	int a;
+	long int a;
 
 	if (!h || !*h)
 		return (0);
